+++ src/index.ts
@@ -1,5 +1,6 @@
 import search from "./search";
 import random from "./random";
+import getAllCombos from "./get-all-combos";
 import findById from "./find-by-id";
 import makeFakeCombo from "./make-fake-combo";
 import { resetCache } from "./spellbook-api";
@@ -7,6 +8,7 @@ import { resetCache } from "./spellbook-api";
 export = {
   search,
   random,
+  getAllCombos,
   findById,
   makeFakeCombo,
   resetCache,